refactor(utils,assets): rename middleware module to utils/core.py (HF-25)
Adopts main's #967 rename on proto, treating the module as the Core Module in
Evans' DDD sense: utils/middleware.py becomes utils/core.py and the companion
MIDDLEWARE_DOMAIN_PATH constant becomes CORE_DOMAIN_PATH.

Verified the module bodies are otherwise identical between branches; adopted
main's "Tiferet Utils Core" docstring title but not its stray double-space
typo. The test module is renamed to tests/utils/test_core.py to match main.

Import-path visible: tiferet.utils.middleware no longer exists. The three
middleware services' DI module_path resolves to tiferet.utils.core, exercised
end-to-end by the existing cache_middleware resolution test.

Reclassified from FR-2's sibling FR-1 to a Hotfix RFP: main is canonical,
the change is mechanical, and no design decision is open.
