+++ tiferet/utils/__init__.py
@@ -29,4 +29,4 @@ from .yaml import YamlLoader, YamlLoader as Yaml
 from .toml import TomlLoader, TomlLoader as Toml
 from .csv import CsvLoader, CsvLoader as Csv, CsvDictLoader, CsvDictLoader as CsvDict
 from .sqlite import SqliteClient, SqliteClient as Sqlite
-from .middleware import LoggingMiddleware, TimingMiddleware, CacheMiddleware
+from .core import LoggingMiddleware, TimingMiddleware, CacheMiddleware
